refactor: move RichText to its own folder
