+++ packages/website/components/ContentfulRichText/index.ts
@@ -1,0 +1,2 @@
+export { RichText } from './RichText';
+export type { RichTextProps } from './RichText';
