+++ packages/website/components/PageContent/index.ts
@@ -1,2 +1,3 @@
 export { PageContent } from './PageContent';
+export type { PageContentProps } from './PageContent';
 export type { HeadingType } from './TableOfContent';
